Added debug routers from resources to debug under router folder
