+++ fhir-backend-dynamic/app/main.py
@@ -1,7 +1,7 @@
 # app/main.py - Updated to include startup system and configuration
 from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
-from app.routers import resources, health, servers, filters, metadata
+from app.routers import resources, health, servers, filters, metadata, debug
 # , aggregate  # TEMPORARILY DISABLED
 from app.core.logging import setup_logging
 from app.startup import initialize_backend, get_startup_status
@@ -17,8 +17,9 @@ app = FastAPI(
     version="2.0.0",
     docs_url="/docs" if ENABLE_DOCS else None,
     redoc_url=None,
-    openapi_url=None,
+    #openapi_url=None,
 )
+
 
 app.add_middleware(
     CORSMiddleware,
@@ -85,6 +86,7 @@ app.include_router(resources.router,  prefix="/api")
 app.include_router(servers.router,    prefix="/api")
 app.include_router(filters.router,    prefix="/api")  # Filter endpoints
 app.include_router(metadata.router,   prefix="/api")  # NEW: FHIR Metadata endpoints
+app.include_router(debug.router,      prefix="/api")
 
 # Conditionally include aggregate router behind feature flag
 # TEMPORARILY DISABLED - causing issues
